requirements.txt
django==5.2
django-debug-toolbar==5.1.0
psycopg2-binary==2.9.10
djangorestframework==3.16.0
drf-nested-routers==0.94.1